LeetCode/tst/com/leetcode/problem139/wordbreak/SolutionTest.java
package com.leetcode.problem139.wordbreak;

import static org.junit.Assert.assertEquals;

import java.util.HashSet;
import java.util.Set;

import org.junit.Before;
import org.junit.Test;

/**
 * Given a string s and a dictionary of words dict, determine if s can be
 * segmented into a space-separated sequence of one or more dictionary words.
 * 
 * For example, given s = "leetcode", dict = ["leet", "code"].
 * 
 * Return true because "leetcode" can be segmented as "leet code".
 * 
 * @author dongyuxi
 *
 */
public class SolutionTest {
    public Solution solution = null;

    @Before
    public void setup() {
        solution = new Solution();
    }

    @Test
    public void test() {
        Set<String> set = new HashSet<String>();
        set.add("leet");
        set.add("code");
        assertEquals(true, solution.wordBreak("leetcode", set));
    }
}